+++ duckscript_sdk/src/sdk/std/json/mod.rs
@@ -1,3 +1,4 @@
+mod encode;
 mod parse;
 
 use crate::utils::pckg;
@@ -6,9 +7,12 @@ use duckscript::types::error::ScriptError;
 
 static PACKAGE: &str = "json";
 
+pub(crate) static OBJECT_VALUE: &str = "[OBJECT]";
+
 pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
     let package = pckg::concat(parent, PACKAGE);
 
+    commands.set(encode::create(&package))?;
     commands.set(parse::create(&package))?;
 
     Ok(())