+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * New json_encode command #124
 * New json_parse command #124
+* Module level documentation #125
+* test_file command is now documented and public.
 
 ### v0.6.3 (2020-07-24)